Move bytecode-related code to separate modules
