+++ src/som/interpreter/bc/bytecodes.py
