+++ src/som/interpreter/bc/frame.py
