+++ src/som/interpreter/bc/interpreter.py
@@ -1,4 +1,4 @@
-from som.interpreter.bytecodes import bytecode_length, Bytecodes
+from som.interpreter.bc.bytecodes import bytecode_length, Bytecodes
 from som.interpreter.control_flow import ReturnException
 
 from rpython.rlib import jit
@@ -275,7 +275,7 @@ class Interpreter(object):
 
 def get_printable_location(bytecode_index, interp, method):
     from som.vmobjects.method import Method
-    from som.interpreter.bytecodes import bytecode_as_str
+    from som.interpreter.bc.bytecodes import bytecode_as_str
     assert isinstance(method, Method)
     bc = method.get_bytecode(bytecode_index)
     return "%s @ %d in %s" % (bytecode_as_str(bc),
